move base worker manager classes into a separate file
