+++ src/typegate/src/runtimes/substantial/workflow_worker_manager.ts
@@ -5,6 +5,11 @@ import { envSharedWithWorkers } from "../../config/shared.ts";
 import { getLogger } from "../../log.ts";
 import { TaskContext } from "../deno/shared_types.ts";
 import {
+  BaseWorker,
+  BaseWorkerManager,
+  DenoWorker,
+} from "../utils/worker_manager.ts";
+import {
   Err,
   Msg,
   Result,
@@ -99,75 +104,6 @@ export class WorkflowRecorder {
     }
 
     return false;
-  }
-}
-
-export abstract class BaseWorker {
-  abstract listen(handlerFn: WorkerEventHandler): void;
-  abstract trigger(type: WorkerEvent, data: unknown): void;
-  abstract destroy(): void;
-  abstract get id(): RunId;
-}
-
-class DenoWorker {
-  #worker: Worker;
-  #runId: RunId;
-  constructor(runId: RunId, workerPath: string) {
-    this.#worker = new Worker(workerPath, {
-      name: runId,
-      type: "module",
-      deno: {
-        permissions: {
-          net: true,
-          // on request permissions
-          read: "inherit", // default read permission
-          sys: "inherit",
-          // non-overridable permissions (security between typegraphs)
-          run: false,
-          write: false,
-          ffi: false,
-          env: envSharedWithWorkers,
-        },
-      },
-    });
-    this.#runId = runId;
-  }
-
-  listen(handlerFn: WorkerEventHandler) {
-    this.#worker.onmessage = async (message) => {
-      if (message.data.error) {
-        // worker level failure
-        await handlerFn(Err(message.data.error));
-      } else {
-        // logic level Result (Ok | Err)
-        await handlerFn(message.data as Result<unknown>);
-      }
-    };
-
-    this.#worker.onerror = /*async*/ (event) => handlerFn(Err(event));
-  }
-
-  trigger(type: WorkerEvent, data: unknown) {
-    this.#worker.postMessage(Msg(type, data));
-  }
-
-  destroy() {
-    this.#worker.terminate();
-  }
-
-  get id() {
-    return this.#runId;
-  }
-}
-
-export class BaseWorkerManager {
-  #workerFactory: (runId: RunId) => BaseWorker;
-  protected constructor(workerFactory: (runId: RunId) => BaseWorker) {
-    this.#workerFactory = workerFactory;
-  }
-
-  get workerFactory() {
-    return this.#workerFactory;
   }
 }
